+++ backend/expenses/Touir.ExpensesManager.Expenses/Infrastructure/Contracts/IRateProvider.cs
@@ -1,4 +1,4 @@
-namespace Touir.ExpensesManager.Expenses.Services
+namespace Touir.ExpensesManager.Expenses.Infrastructure.Contracts
 {
     public interface IRateProvider
     {